Move kernel related tasks in sysctl.yml to sysctl-kernel.yml.
I want to have another set of sysctl options for TCP/IP stack
hardening, which will be put in a sysctl-tcp_ip_stack.yml file.
Then, both the sysctl-*.yml files will be included in the main
sysctl.yml.

Also, introduce the 'security_kern_go_hardcore' variable, which
controls whether the task will be imported or not. This way the user
can choose if he wants to perform the kernel-hardening part.