Completely move to Github Actions from Travis CI.
Travis CI no longer has  access to this repository.
Both the testing and the release of the role to Ansible
Galaxy will be done through Github Actions workflows.

I might use Travis in the future for testing with TOX, but
for now GitHub actions will be the only testing environment.