Fix #3: Configure pam_tally2 only on older Archlinux systems.
Archlinux is moving from pam_tally2 to pam_faillock, just like
RedHat-based distributions. The issue was that in latest
Archlinux systems this role would configure pam_tally2.so,
resulting in both pam_tally2.so and pam_faillock.so being used,
and therefore locking the user out of the system.

With this commit, pam_tally2.so configuration will be skipped
if pam_faillock is being used,

Also, use the pamd module instead of replace or lineinfile.