Add the "ignoreip" fail2ban option.
This option allows the user to specify a list of IPs to be
whitelisted, meaning, fail2ban will not ban a host which
matches an address on this list. This can be extremely
useful when using an IP banning software like fail2ban,
since it makes sure than if an attacker ever finds out the
user's IP address, they cannot send packets with spoofed
headers to get the user banned from their own server(s).